+++ src/interfaces/IeETH.sol
@@ -34,7 +34,7 @@ interface IeETH {
 
     function pause() external;
     function pauseUntil(address _user) external;
-    function extendPauseUntil(address _user, uint64 _duration) external;
+    function extendPauseUntil(address _user, uint256 _duration) external;
     function cancelPauseUntil(address _user) external;
     function unpause() external;
